requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.7.1
mongoengine==0.27.0
python-dotenv==1.0.0
pytest==7.4.0
gunicorn==21.2.0
marshmallow==3.20.1
